Combine Ghostwriter context selectors into one dropdown
